Add SearchQueryForm componentg
